Kyle_Model_V3/requirements.txt
# 强化学习相关
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# 数据处理和分析
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
tqdm>=4.62.0
jupyter>=1.0.0
ipython>=7.0.0
PyYAML>=6.0.0

# 可选：更高级的可视化
# plotly>=5.0.0
# bokeh>=2.4.0

# 可选：多智能体环境（如果需要扩展）
# pettingzoo>=1.18.0
# ray[rllib]>=2.0.0